Move to fixit.cli and backwards compat
